% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_id_from_url.R
\name{album_id_from_url}
\alias{album_id_from_url}
\title{Extract Album URL ID}
\usage{
album_id_from_url(url)
}
\arguments{
\item{url}{A valid URL}
}
\value{
Returns a glm object containing the trained MIMoSA coefficients.
}
\description{
From the URL information extract each album ID of an artist
}
\examples{
\dontrun{
URL = "https://genius.com/albums/Game-of-thrones/Season-1-scripts"
info = album_id_from_url(url = URL)
}
}
