% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanGoT.R
\name{cleanGoT}
\alias{cleanGoT}
\title{Clean GoT Script Data}
\usage{
cleanGoT(data)
}
\arguments{
\item{data}{The data returned by scraping using geniusr}
}
\value{
Returns a semi cleaned dataset
}
\description{
Cleans GoT data taken from geniusr scrape
}
\examples{
\dontrun{
base_url = "https://genius.com/albums/Game-of-thrones"
info = GoT::scrape_GoT(base_url = base_url, season = 1)
cleaned = GoT::cleanGoT(info)
}
}
