% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time.com_deaths.R
\name{get_time.com_deaths}
\alias{get_time.com_deaths}
\title{GoT Deaths and Times from time.com}
\usage{
get_time.com_deaths(url)
}
\arguments{
\item{url}{The time.com url 'http://time.com/3924852/every-game-of-thrones-death/'}
}
\value{
Returns a cleaned dataset of the deaths in GoT as a tibble
}
\description{
Obtains GoT character deaths and times throughout series
}
\examples{
\dontrun{
url = 'http://time.com/3924852/every-game-of-thrones-death/'
time.com_deaths  = get_time.com_deaths()
}
}
