% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_from_url.R
\name{info_from_url}
\alias{info_from_url}
\title{Extract Information from URL}
\usage{
info_from_url(url)
}
\arguments{
\item{url}{A valid URL}
}
\value{
Returns URL info
}
\description{
Obtain information from URL like multiple albums and IDs
}
\examples{
\dontrun{
URL = "https://genius.com/albums/Game-of-thrones/Season-1-scripts"
info = info_from_url(url = URL)
}
}
