% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parens_to_colon.R
\name{parens_to_colon}
\alias{parens_to_colon}
\title{Force colon}
\usage{
parens_to_colon(x)
}
\arguments{
\item{x}{A string}
}
\value{
Returns a string with a colon
}
\description{
Change lines without a colon to have a colon after character name
}
\examples{
\dontrun{
x = "Tywin (to Sansa) Stupid girl"
parens_to_colon(x=x)
}
}
