% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_characters.R
\name{scrape_characters}
\alias{scrape_characters}
\title{Extract Character Names From Wiki}
\usage{
scrape_characters(url)
}
\arguments{
\item{url}{A valid URL}
}
\value{
A tibble of character names and description
}
\description{
Obtain character names from http://awoiaf.westeros.org/index.php/List_of_characters
}
\examples{
\dontrun{
url = "http://awoiaf.westeros.org/index.php/List_of_characters"
characters = scrape_characters(url)
}
}
