% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_full_scripts.R
\name{scrape_full_scripts}
\alias{scrape_full_scripts}
\title{Scrape GoT Full Script Data}
\usage{
scrape_full_scripts(song_lyrics_url, access_token = genius_token())
}
\arguments{
\item{song_lyrics_url}{url for data to scrape}

\item{access_token}{The API token}
}
\value{
Returns the dataset pulled from scraping
}
\description{
Scrapes and episode of GoT given the URL. Copied and adapted from geniusr package scrape_lyrics_url function
}
\examples{
\dontrun{
url = "https://genius.com/albums/Game-of-thrones"
script = scrape_full_scripts(url)
}
}
