% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_GoT.R
\name{scrape_GoT}
\alias{scrape_GoT}
\title{Scrape GoT Data}
\usage{
scrape_GoT(base_url, seasons)
}
\arguments{
\item{base_url}{The base URL for all seasons of GoT}

\item{seasons}{The season you would like to scrape data up until}
}
\value{
Returns the dataset pulled from genius using geniusr
}
\description{
Scrapes all episodes for all seasons of GoT
}
\examples{
\dontrun{
base_url = "https://genius.com/albums/Game-of-thrones"
info = scrape_GoT(base_url = base_url, season = 1)
}
}
