% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_HBO_characters.R
\name{scrape_HBO_characters}
\alias{scrape_HBO_characters}
\title{Extract Character Names From Wiki}
\usage{
scrape_HBO_characters(url)
}
\arguments{
\item{url}{A valid URL}
}
\value{
Returns a cleaned dataset of the character names from HBO
}
\description{
Obtain character names from https://www.hbo.com/game-of-thrones/cast-and-crew
}
\examples{
\dontrun{
url = 'https://www.hbo.com/game-of-thrones/cast-and-crew'
hbo_characters = scrape_HBO_characters(url)
}
}
