% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_death_times.R
\name{get_death_times}
\alias{get_death_times}
\title{GoT Deaths and Times}
\usage{
get_death_times()
}
\arguments{
\item{filename}{path to store html file extracted}
}
\value{
Returns a cleaned dataset of the deaths in GoT as a tibble
}
\description{
Obtains GoT character deaths and times throughout series
}
\examples{
\dontrun{
filename = "frozen_death_times.html"
cleaned_death = get_death_times(filename = filename)
}
}
